consolidate fetch helpers into api file
