+++ components/app.tsx
@@ -6,7 +6,7 @@ import HourView from '@/components/hour-view'
 import ChartView from '@/components/chart-view'
 import TableView from '@/components/table-view'
 import { getDayEnd, getPrevWeek } from '@/lib/date-util'
-import { postBody, handleEntryResponse } from '@/lib/fetch-util'
+import { postBody, handleEntryResponse } from '@/lib/api'
 import styles from '@/styles/App.module.css'
 
 type AppProps = {
