+++ lib/api.ts
@@ -1,7 +1,35 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { getSession } from 'next-auth/react'
-import type { EntryResponse } from '@/lib/types'
+import type { EntryData, EntryResponse } from '@/lib/types'
 import { isEntryData } from '@/lib/types'
+
+const postBody = (data: object) => {
+    return {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(data)
+    }
+}
+
+const handleEntryResponse = async (res: Response): Promise<Array<EntryData>> => {
+    const entryResponse: EntryResponse = await res.json()
+    // check and log errors
+    if (!res.ok) {
+        const message = 'message' in entryResponse
+            ? entryResponse.message
+            : 'failed'
+        throw new Error(`Api Error ${res.status}: ${message}`)
+    }
+    // return empty list of entries if success and message
+    if ('message' in entryResponse) {
+        return []
+    }
+    // parse json string dates if entries returned
+    return entryResponse.data.map((entry: EntryData) => {
+        entry.date = new Date(entry.date)
+        return entry
+    })
+}
 
 const hasSession = async (req: NextApiRequest, res: NextApiResponse<EntryResponse>): Promise<boolean> => {
     const session = await getSession({ req })
@@ -58,5 +86,7 @@ export {
     hasEmail,
     hasTimeBounds,
     hasEntry,
-    hasIds
+    hasIds,
+    handleEntryResponse,
+    postBody
 }
